recommender_system/requirements.txt
tqdm==4.64.1
python-dotenv==0.21.1
loguru==0.6.0
numpy==1.24.1
motor==3.1.1
backoff==2.1.2
python-dateutil==2.4
